+++ Wallets/Http/Resources/WalletResource.php
@@ -21,12 +21,12 @@ class WalletResource extends JsonResource
         $totalTransfer = number_format(abs($this->transactions()->where('wallet_id',$this->id)->where('meta->type', 'Transfer')->sum('amount')) / 100 ,2);
         return [
             'name' => $this->name,
-            'balance' => number_format($this->balanceFloat,2),
-            'transactions_count' => $this->transactions->count(),
-            'transfers_count' => $this->transfers->count(),
-            'total_transfer' => $totalTransfer,
-            'total_received' => $totalReceived,
-            'total_spent' => $totalSpent
+            'balance' => (float) number_format($this->balanceFloat,2),
+            'transactions_count' => (int) $this->transactions->count(),
+            'transfers_count' => (int) $this->transfers->count(),
+            'total_transfer' => (int) $totalTransfer,
+            'total_received' => (int) $totalReceived,
+            'total_spent' => (int) $totalSpent
         ];
     }
 }
